printf with some categories ok

+++ printf_home/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: gsims <marvin@42lausanne.ch>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/19 16:33:07 by gsims             #+#    #+#             */
-/*   Updated: 2023/10/24 15:55:46 by gsims            ###   ########.fr       */
+/*   Updated: 2023/10/24 17:21:50 by gsims            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 # include <stdarg.h>
 
 int		ft_printchar(int num);
-void	ft_printchar_rec(int num);
+void	ft_printchar_rec(unsigned int num);
 int		ft_printstr(char *s);
 int		ft_printint(int num);
 int		ft_printunsigned(unsigned int num);
@@ -35,6 +35,6 @@ void	ft_putptr(unsigned long long num);
 int		ft_char(char c);
 int		ft_printf(const char *s, ...);
 int		ft_countint(int num);
-int		ft_countint_hex(unsigned int num);
+int		ft_countint_hex(unsigned long long num);
 
 #endif
